feat: skip if the publiccode.yml was updated
Don't open or update the issue if the publiccode.yml was updated
after the log date, but instead skip the repo and postpone the check to
the next run (usually after 24 hours).

If the publiccode.yml has changed it means the maintainer did potentially
made our error obsolete, or might have fixed all the errors, making the
publiccode.yml valid, and at that point it's fair for them to close the
issue.

Without this check, we'd be annoyingly reopen an obsolete issue.